benchmark data: remove results for >12 domains
Our analysis of the moonpool implementation shows that it 'clamps' the
number of domains to `Domain.recommended_domain_count ()` in any case,
so the moonpool results for DOMAINS=15 (for example) are in fact
running with only 12 domains. This easily explains our observation
that Moonpool scales in CPU-contended scenarios, and in fact it
nullifies it. Given that we did not have any interesting qualitative
observation to make on these CPU-contended results, they do not bring
anything much (except validating the fact that runtime performance
degrade sharply), and it is simpler to just remove them from the
plots.